EnvironmentVariables added

+++ PrescriptionUpload/src/com/ADA/AdvancedHealthAcademy/Service/PrescriptionsService.java
@@ -5,6 +5,7 @@ import com.ADA.AdvancedHealthAcademy.DTO.PrescriptionInfoDTO;
 import com.ADA.AdvancedHealthAcademy.Entity.Prescriptions;
 import com.ADA.AdvancedHealthAcademy.Exceptions.ResourceNotFoundException;
 import com.ADA.AdvancedHealthAcademy.Repository.PrescriptionsRepository;
+import com.ADA.AdvancedHealthAcademy.Repository.PrescriptionsUploadRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.ResponseEntity;
@@ -19,31 +20,63 @@ public class PrescriptionsService {
     private PrescriptionsRepository prescriptionsRepository;
     @Autowired
     private PrescriptionConverter prescriptionConverter;
+    @Autowired
+    private PrescriptionsUploadRepository prescriptionsUploadRepository;
 
     public List<PrescriptionInfoDTO> findAllPrescriptionAndMedicineDataByProfileId(long profileId) throws Exception {
         try {
-        List<Object[]> results = prescriptionsRepository.findAllPrescriptionAndMedicineDataByProfileId(profileId);
-        List<PrescriptionInfoDTO> prescriptionInfoDTOS = prescriptionConverter.convertQueryToPrescriptionInfoDTO(results);
-        return prescriptionInfoDTOS;
+            Long count = prescriptionsUploadRepository.profileCountById(profileId);
+            if(count==1){
+                boolean is_active = prescriptionsUploadRepository.isActiveProfile(profileId);
+                if(is_active){
+                    List<Object[]> results = prescriptionsRepository.findAllPrescriptionAndMedicineDataByProfileId(profileId);
+                    List<PrescriptionInfoDTO> prescriptionInfoDTOS = prescriptionConverter.convertQueryToPrescriptionInfoDTO(results);
+                    return prescriptionInfoDTOS;
+                }else {
+                    throw new Exception("Profile is Inactive with ID: "+ profileId);
+                }
+            }else {
+                throw new Exception("Profile not found with ID: "+ profileId);
+            }
         } catch (DataAccessException e) {
-            throw new Exception("Error retrieving all prescriptions by profileId: " + e.getMessage());
+            throw new Exception("Error retrieving all prescriptions by ProfileId: " + e.getMessage());
         }
     }
 
 
     public Optional<PrescriptionInfoDTO> findPrescriptionById(Long profileId,Long prescriptionId) throws Exception {
         try {
-            List<Object[]> results = prescriptionsRepository.findPrescriptionAndMedicineByProfileIdAndPrescriptionId(profileId,prescriptionId);
-            Optional<PrescriptionInfoDTO> prescriptionInfoDTO = prescriptionConverter.convertPrescriptionAndMedicine(results);
-            return prescriptionInfoDTO;
+            Long count = prescriptionsUploadRepository.profileCountById(profileId);
+            if(count==1){
+                boolean is_active = prescriptionsUploadRepository.isActiveProfile(profileId);
+                if(is_active){
+                    List<Object[]> results = prescriptionsRepository.findPrescriptionAndMedicineByProfileIdAndPrescriptionId(profileId,prescriptionId);
+                    Optional<PrescriptionInfoDTO> prescriptionInfoDTO = prescriptionConverter.convertPrescriptionAndMedicine(results);
+                    return prescriptionInfoDTO;
+                }else {
+                    throw new Exception("Profile is Inactive with ID: "+ profileId);
+                }
+            }else {
+                throw new Exception("Profile not found with ID: "+ profileId);
+            }
         } catch (DataAccessException e) {
             throw new Exception("Error finding prescription by ID: " + e.getMessage());
         }
     }
 
-    public void deletePrescriptionById(Long prescriptionId) throws Exception {
+    public void deletePrescriptionById(Long profileId,Long prescriptionId) throws Exception {
         try {
-            prescriptionsRepository.deleteById(prescriptionId);
+            Long count = prescriptionsUploadRepository.profileCountById(profileId);
+            if(count==1){
+                boolean is_active = prescriptionsUploadRepository.isActiveProfile(profileId);
+                if(is_active){
+                    prescriptionsRepository.deleteById(prescriptionId);
+                }else {
+                    throw new Exception("Profile is Inactive with ID: "+ profileId);
+                }
+            }else {
+                throw new Exception("Profile not found with ID: "+ profileId);
+            }
         } catch (DataAccessException e) {
             throw new Exception("Error deleting prescription: " + e.getMessage());
         }
@@ -51,14 +84,25 @@ public class PrescriptionsService {
 
     public String deleteAllPrescriptionByIdProfileId(Long profileId) throws Exception {
         try {
-            List<Prescriptions> prescriptionOptional = prescriptionsRepository.findAllPrescriptionByProfileId(profileId);
-            if (!prescriptionOptional.isEmpty()) {
-                for (int k=0;k<=prescriptionOptional.size()-1;k++){
-                   prescriptionsRepository.deleteById(prescriptionOptional.get(k).getPrescriptionId()); ;
+            Long count = prescriptionsUploadRepository.profileCountById(profileId);
+            if(count==1){
+                boolean is_active = prescriptionsUploadRepository.isActiveProfile(profileId);
+                if(is_active){
+
+                    List<Prescriptions> prescriptionOptional = prescriptionsRepository.findAllPrescriptionByProfileId(profileId);
+                    if (!prescriptionOptional.isEmpty()) {
+                        for (int k=0;k<=prescriptionOptional.size()-1;k++){
+                            prescriptionsRepository.deleteById(prescriptionOptional.get(k).getPrescriptionId()); ;
+                        }
+                        return "Successfully Deleted All prescriptions";
+                    }else {
+                        throw new ResourceNotFoundException("Prescription Does not exist with profile ID: " + profileId);
+                    }
+                }else {
+                    throw new Exception("Profile is Inactive with ID: "+ profileId);
                 }
-                return "Successfully Deleted All prescriptions";
             }else {
-                throw new ResourceNotFoundException("Prescription Does not exist with profile ID: " + profileId);
+                throw new Exception("Profile not found with ID: "+ profileId);
             }
         } catch (DataAccessException e) {
             throw new Exception("Error deleting All prescriptions by profile Id: "+ profileId +" "+ e.getMessage());
